+++ src/bpp/models/autor.py
@@ -78,16 +78,6 @@ class Autor(LinkDoPBNMixin, ModelZAdnotacjami, ModelZPBN_ID):
     oraz na podstronie autora, po nazwisku i tytule naukowym.""",
     )
 
-    aktualny = models.BooleanField(
-        "Aktualny?",
-        default=False,
-        help_text="""Jeżeli zaznaczone, pole to oznacza,
-    że autor jest aktualnie - na dziś dzień - przypisany do jakiejś jednostki w bazie danych i jego przypisanie
-    do tej jednostki nie zostało zakończone wraz z konkretną datą w
-    przeszłości.""",
-        db_index=True,
-    )
-
     aktualna_jednostka = models.ForeignKey(
         "Jednostka", CASCADE, blank=True, null=True, related_name="aktualna_jednostka"
     )